+++ Examples/selection.sh
@@ -44,9 +44,9 @@ sparkle run solvers --performance-data
 
 sparkle jobs
 
-# Now we can construct a portfolio selector, using the previously computed features and the objective value results of running the solvers. We can specify an objective to select on with the `--objective` flag, but if we do not, Sparkle will default to the first objective specified in the Settings file. The `--selector-timeout` argument determines for how many seconds we will train our selector for. We can set the flag `--solver-ablation` for actual marginal contribution computation later.
+# Now we can construct a portfolio selector, using the previously computed features and the objective value results of running the solvers. We can specify an objective to select on with the `--objective` flag, but if we do not, Sparkle will default to the first objective specified in the Settings file. We can set the flag `--solver-ablation` for actual marginal contribution computation later.
 
-sparkle construct portfolio selector --selector-timeout 1000 --solver-ablation
+sparkle construct portfolio selector --solver-ablation
 sparkle jobs  # Wait for the constructor to complete its computations
 
 ### Generate a report
@@ -106,7 +106,7 @@ sparkle jobs
 
 # Now we can train a selector based on these features.
 
-sparkle construct portfolio selector --selector-timeout 1000
+sparkle construct portfolio selector 1000
 sparkle jobs  # Wait for the computation to be done
 
 # And generate the report. When running on the PTN/PTN2 data sets, you can compare the two to see the impact of different feature extractors.